noir_ssa

Module cli

Modules

interpret_cmd 🔒

transform_cmd 🔒

visualize_cmd 🔒

Structs

SsaArgs 🔒

Common SSA command parameters.

SsaCli 🔒

Enums

SsaCommand 🔒

Constants

PKG_VERSION 🔒

Statics

VERSION_STRING 🔒

Functions

parse_ssa 🔒

Parse the SSA.

read_source 🔒

Read the SSA from a file or stdin.

ssa_passes 🔒

List of the SSA passes in the primary pipeline, enriched with their “step” count so we can use unambiguous naming in filtering.

start_cli 🔒

write_output 🔒

Write the output to a file or stdout.