noir_ssa::cli
Function parse_ssa

fn parse_ssa(src: &str, validate: bool) -> Result<Ssa>

Parse the SSA.

If parsing fails, print errors to stderr and return a failure.

If validation is enabled, any semantic error causes a panic.