noir_ssa::cli

Module interpret_cmd

Structs

InterpretCommand 🔒

Parse the input SSA and it arguments, run the SSA interpreter, then write the return values to stdout.

Constants

TOML_LINE_SEP 🔒

Functions

abi_from_ssa 🔒

Derive an ABI description from the SSA parameters.

abi_type_from_multi_ssa 🔒

Create an ABI type from multiple SSA types, for example when multiple values are returned, or appear in arrays.

abi_type_from_ssa 🔒

Create an ABI type from a single SSA type.

read_inputs_and_return 🔒

run 🔒

write_to_temp_file 🔒